Add further abstraction.
